Core-Styles: WIP: Allow Script Usage Not Just CLI
