+++ bin/config.js
@@ -1,3 +1,5 @@
+#!/usr/bin/env node
+
 const fs = require('fs');
 const merge = require('merge-lite');
 const yaml = require('js-yaml');
